Updated ListBackend

+++ Ubyelp/src/components/BusinessList/ListBackend.js
@@ -41,6 +41,7 @@ class ListBackend{
         this.disconnect = this.disconnect.bind(this);
 
         this.#queryVerifyName;
+        this.#queryVerifyListItem;
         this.#queryGenerateListId;
         this.#queryGenerateListItemId;
         this.#quickQuery;
@@ -51,8 +52,8 @@ class ListBackend{
     //Create List (Requires: name)
     async createList(name, desc) {
         //Check if name exists
-        if (name == '') throw 'List name cannot be empty!'
-        await this.#queryVerifyName(name);
+        if (name == '') return 'List name cannot be empty!'
+        if(await this.#queryVerifyName(name)) return 'Name already exists!';
 
         //Generate list id
         var list_id = await this.#queryGenerateListId();
@@ -66,12 +67,13 @@ class ListBackend{
         await this.#quickQuery(query);
 
         console.log('List ' + list_id + ' created');
+        return '';
     }
 
     //Remove list
     async removeList(list_id) {
         //Remove list
-        await this.#queryAccountListAccess(list_id);
+        if(!await this.#queryAccountListAccess(list_id)) return 'Illegal access to list';
 
         var query = 'DELETE FROM ' + list_table + ' WHERE `list_id`=' + list_id;
         await this.#quickQuery(query);
@@ -81,30 +83,36 @@ class ListBackend{
         await this.#quickQuery(query);
 
         console.log('List ' + list_id + ' removed');
+        return '';
     }
 
     //Edit title of List
     async editListName(list_id, newTitle) {
-        await this.#queryAccountListAccess(list_id);
+        if (!await this.#queryAccountListAccess(list_id)) return 'Illegal access to list';
 
         var query = 'UPDATE ' + list_table + ' SET `name`=\'' + newTitle + '\' WHERE `list_id`=' + list_id;
         await this.#quickQuery(query);
         console.log('List name changed');
+        return '';
     }
 
     //Edit description of list
     async editListDescription(list_id, newDesc) {
-        await this.#queryAccountListAccess(list_id);
+        if (!await this.#queryAccountListAccess(list_id)) return 'Illegal access to list';
 
         const query = 'UPDATE ' + list_table + ' SET `desc`=\'' + newDesc + '\' WHERE `list_id`=' + list_id;
         await this.#quickQuery(query);
         console.log('List description changed');
+        return '';
     }
 
     //Add list item (Requires: link)
     async addListItem(list_id, name, link, desc, rating) {
-        if (link == '') throw 'Link not provided';
-        await this.#queryAccountListAccess(list_id);
+        if (link == '') return 'Link not provided';
+        if (!await this.#queryAccountListAccess(list_id)) return 'Illegal access to list!';
+
+        //Check if item already added
+        if (await this.#queryVerifyListItem(list_id,link)) return '';
 
         //Generate list item id
         let list_item_id = await this.#queryGenerateListItemId();
@@ -118,11 +126,12 @@ class ListBackend{
         await this.#quickQuery(query);
 
         console.log('List item added');
+        return '';
     }
 
     //Remove list item from list
     async removeListItem(list_id, list_item_id) {
-        await this.#queryAccountListItemAccess(list_item_id);
+        if(!await this.#queryAccountListItemAccess(list_item_id)) return 'Illegal access to list!';
 
         //Remove list_item
         var query = 'DELETE FROM ' + list_item_table + ' WHERE `list_item_id`=' + list_item_id;
@@ -133,47 +142,52 @@ class ListBackend{
         await this.#quickQuery(query);
 
         console.log('List Item ' + list_item_id + ' removed');
+        return '';
     }
 
     //Add/Edit description to list item
     async editItemDescription(list_item_id, newDesc) {
-        await this.#queryAccountListItemAccess(list_item_id);
+        if (!await this.#queryAccountListItemAccess(list_item_id)) return 'Illegal access to list!';
 
         const query = 'UPDATE ' + list_item_table + ' SET `desc`=\'' + newDesc + '\' WHERE `list_item_id`=' + list_item_id;
         await this.#quickQuery(query);
 
         console.log('List item description change');
+        return '';
     }
 
     //Rate list item
     async editItemRating(list_item_id, rating) {
-        await this.#queryAccountListItemAccess(list_item_id);
+        if (!await this.#queryAccountListItemAccess(list_item_id)) return 'Illegal access to list!';
 
         const query = 'UPDATE ' + list_item_table + ' SET `rating`=\'' + rating + '\' WHERE `list_item_id`=' + list_item_id;
         await this.#quickQuery(query);
 
         console.log('List item rating changed');
+        return '';
     }
 
     //Change list item name
     async editItemName(list_item_id, name) {
-        await this.#queryAccountListItemAccess(list_item_id);
+        if (!await this.#queryAccountListItemAccess(list_item_id)) return 'Illegal access to list!';
 
         const query = 'UPDATE ' + list_item_table + ' SET `name`=\'' + name + '\' WHERE `list_item_id`=' + list_item_id;
         await this.#quickQuery(query);
 
         console.log('List item name changed');
+        return '';
     }
 
     //Get all non-empty lists from Database
     async getLists(callback) {
-        const query = 'SELECT list.name AS \'listname\', list.desc AS \'listdesc\', list_item.name AS \'itemname\', list_item.desc AS \'itemdesc\', list_item.rating AS \'rating\', list_item.link AS \'link\' FROM ' + account_list_relation + ' JOIN ' + list_item_relation + ' USING(list_id) JOIN ' + list_table + ' USING(list_id) JOIN ' + list_item_table + ' USING(list_item_id) WHERE `account_id`=' + this.account_id;
+        const query = 'SELECT list_id, list_item_id, list.name AS \'listname\', list.desc AS \'listdesc\', list_item.name AS \'itemname\', list_item.desc AS \'itemdesc\', list_item.rating AS \'rating\', list_item.link AS \'link\' FROM ' + account_list_relation + ' JOIN ' + list_item_relation + ' USING(list_id) JOIN ' + list_table + ' USING(list_id) JOIN ' + list_item_table + ' USING(list_item_id) WHERE `account_id`=' + this.account_id;
         this.Connection.query(query, function (err, rows, fields) {
             if (err) callback(err, null);
 
             let list = {};
             for (var i = 0; i < rows.length; i++) {
                 let listitem = {
+                    'id': rows[i].list_item_id,
                     'name': rows[i].itemname,
                     'desc': rows[i].itemdesc,
                     'rating': rows[i].rating,
@@ -181,6 +195,7 @@ class ListBackend{
                 };
                 if (!list[rows[i].listname]) {
                     list[rows[i].listname] = {
+                        'id': rows[i].list_id,
                         'desc': rows[i].listdesc,
                         'items': [listitem]
                     }
@@ -199,14 +214,28 @@ class ListBackend{
     /*Helper functions*/
     #queryVerifyName(name) {
         return new Promise((resolve, reject) => {
-            const query = 'SELECT name FROM ' + list_table + ' JOIN ' + account_list_relation + ' USING(list_id);';
+            const query = 'SELECT name FROM ' + list_table + ' JOIN ' + account_list_relation + ' USING(list_id) WHERE account_id=' + this.account_id;
             this.Connection.query(query, function (err, rows, fields) {
                 if (err) throw err;
 
                 for (var i = 0; i < rows.length; i++) {
-                    if (name == rows[i].name) reject(new Error('Name already exists!'));
+                    if (name == rows[i].name) resolve(true);
                 }
-                resolve();
+                resolve(false);
+            });
+        });
+    }
+
+    #queryVerifyListItem(list_id, link) {
+        return new Promise((resolve, reject) => {
+            const query = 'SELECT link FROM ' + list_table + ' JOIN ' + account_list_relation + ' USING(list_id) JOIN ' + list_item_relation + ' USING(list_id) JOIN ' + list_item_table + ' USING(list_item_id) WHERE account_id = ' + this.account_id + ' AND list_id = ' + list_id;
+            this.Connection.query(query, function (err, rows, fields) {
+                if (err) throw err;
+
+                for (var i = 0; i < rows.length; i++) {
+                    if (link == rows[i].link) resolve(true);
+                }
+                resolve(false);
             });
         });
     }
@@ -253,9 +282,9 @@ class ListBackend{
                 if (err) reject(err);
 
                 for (var i = 0; i < rows.length; i++) {
-                    if (list_id == rows[i].list_id) resolve();
+                    if (list_id == rows[i].list_id) resolve(true);
                 }
-                reject(new Error('ERROR: Illegal access to list!'));
+                resolve(false);
             });
         });
     }
@@ -267,9 +296,9 @@ class ListBackend{
                 if (err) reject(err);
 
                 for (var i = 0; i < rows.length; i++) {
-                    if (list_item_id == rows[i].list_item_id) resolve();
+                    if (list_item_id == rows[i].list_item_id) resolve(true);
                 }
-                reject(new Error('ERROR: Illegal access to list!'));
+                resolve(false);
             });
         });
     }
@@ -280,10 +309,11 @@ module.exports = ListBackend;
 //Testing
 async function test() {
     var list = new ListBackend(663);
-    await list.createList('new', 'desc');
+    await list.createList('list3', 'desc');
     await list.editListName(1, 'new name 2');
     await list.editListDescription(1, 'new desc');
-    await list.addListItem(1, 'Aple', 'www.yelp.com/asfsdhkgdl', 'local apple store', 4);
+    await list.addListItem(4, 'Aple', 'www.yelp.com/asfsdhkgdl', 'local apple store', 4);
+    await list.addListItem(4, 'Aple', 'www.yelp.com/asfsdhkgdl', 'local apple store', 4);   //Repeat
     await list.editItemDescription(1, 'old apple store');
     await list.editItemRating(1, 1);
     await list.editItemName(1, '3rd Apple Store');
